Create TextSelectCopy component
