+++ packages/shared/components/TextSelectCopy/TextSelectCopy.tsx
@@ -1,0 +1,51 @@
+/**
+ * Copyright 2020 Gravitational, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import React from 'react';
+import copyToClipboard from 'design/utils/copyToClipboard';
+import selectElementContent from 'design/utils/selectElementContent';
+import { ButtonPrimary, Text, Flex } from 'design';
+
+export default function TextSelectCopy({ text }: Props) {
+  const ref = React.useRef();
+  const [copyCmd, setCopyCmd] = React.useState(() => 'Copy');
+
+  function onCopyClick() {
+    copyToClipboard(text).then(() => setCopyCmd('Copied'));
+    selectElementContent(ref.current);
+  }
+
+  return (
+    <Flex
+      bg="bgTerminal"
+      p="2"
+      mb={2}
+      alignItems="center"
+      justifyContent="space-between"
+    >
+      <Text ref={ref} style={{ wordBreak: 'break-all' }} mr="3">
+        {text}
+      </Text>
+      <ButtonPrimary onClick={onCopyClick} size="small">
+        {copyCmd}
+      </ButtonPrimary>
+    </Flex>
+  );
+}
+
+type Props = {
+  text: string;
+};
